+++ src/dao/ListaPeliculas.java
@@ -6,7 +6,7 @@
  * @version: 1.0
  */
 
-package datos;
+package dao;
 
 import java.io.File;
 import java.util.ArrayList;
